refactor(issue-25): S6 fix two preexisting lint errors
1. LanguageContext.jsx — react-refresh/only-export-components: split
   non-component exports out so the .jsx file only exports the
   LanguageProvider component. The context constant and useLanguage
   hook move to useLanguage.js; index.js re-exports both, and
   CodeEditor.jsx updates its import path.

2. useRuntime.js — react-hooks/set-state-in-effect: stop calling
   setWarm(false) and setWarmStats(null) inside the effect. Track
   prevLang via useState and reset both during render when languageId
   changes, following the React-docs "adjusting state when a prop
   changes" pattern.

Lint now exits 0 (3 warnings remain, all preexisting and orthogonal).
Build still succeeds. Manual verification via smoke-test-multilang.mjs:
cpp and python runners both execute correctly.

Refs #25
