+++ apps/frontend/src/widgets/code-editor/index.js
@@ -1,3 +1,4 @@
 export { default as CodeEditor } from './CodeEditor.jsx'
-export { LanguageProvider, useLanguage } from './LanguageContext.jsx'
+export { LanguageProvider } from './LanguageContext.jsx'
+export { useLanguage } from './useLanguage.js'
 export { runtimes, runtimeList, getRuntime } from './runtimes/index.js'
